+++ app/src/main/java/com/sysadmindoc/snapcrop/CollageActivity.kt
@@ -136,6 +136,7 @@ class CollageActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         enableEdgeToEdge()
+        applySecureWindow(this)
         @Suppress("DEPRECATION")
         intent.getParcelableArrayListExtra<Uri>(InboundShareContract.EXTRA_URIS)
             ?.distinctBy(Uri::toString)
@@ -230,6 +231,11 @@ class CollageActivity : ComponentActivity() {
                 }
             }
         }
+    }
+
+    override fun onResume() {
+        super.onResume()
+        applySecureWindow(this)
     }
 
     companion object {